+++ variables.tf
@@ -134,3 +134,9 @@ variable "cluster_delete_timeout" {
   description = "Timeout value when deleting the EKS cluster."
   default     = "15m"
 }
+
+variable "local_exec_interpreter" {
+  description = "Command to run for local-exec resources. Must be a shell-style interpreter. If you are on Windows Git Bash is a good choice. Defaults to [\"/bin/sh\", \"-c\"]"
+  type        = "list"
+  default     = ["/bin/sh", "-c"]
+}